Mobile-Application-dev/Mini_Project_Notes_App/backend/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
passlib
bcrypt
python-jose
pwdlib
PyJWT
python-multipart